Generalized existing behavior for use with both latest & selected posts
